#3 Modelo Activity y Migración

+++ database/migrations/2020_07_01_223903_create_activities_table.php
@@ -17,7 +17,7 @@ class CreateActivitiesTable extends Migration
             $table->bigIncrements('id');
             $table->string('title');
             $table->text('description');
-            $table->integer('score');
+            $table->float('score');
             $table->timestamps();
         });
     }
